Moved get_machine_identifier to salt utils platfrom as per reviewers request
