+++ salt/utils/files.py
@@ -905,18 +905,3 @@ def get_encoding(path):
         return "ASCII"
 
     raise CommandExecutionError("Could not detect file encoding")
-
-
-def get_machine_identifier():
-    """
-    Provide the machine-id for machine/virtualization combination
-    """
-    # Provides:
-    #   machine-id
-    locations = ["/etc/machine-id", "/var/lib/dbus/machine-id"]
-    existing_locations = [loc for loc in locations if os.path.exists(loc)]
-    if not existing_locations:
-        return {}
-    else:
-        with salt.utils.files.fopen(existing_locations[0]) as machineid:
-            return {"machine_id": machineid.read().strip()}
