+++ tests/pytests/functional/utils/gitfs/test_gitfs.py
@@ -2,10 +2,10 @@ import os.path
 
 import pytest
 
-import salt.utils.files
 from salt.fileserver.gitfs import PER_REMOTE_ONLY, PER_REMOTE_OVERRIDES
 from salt.utils.gitfs import GitFS, GitPython, Pygit2
 from salt.utils.immutabletypes import ImmutableDict, ImmutableList
+from salt.utils.platform import get_machine_identifier as _get_machine_identifier
 
 pytestmark = [
     pytest.mark.slow_test,
@@ -249,9 +249,7 @@ def _test_lock(opts):
     g.fetch_remotes()
     assert len(g.remotes) == 1
     repo = g.remotes[0]
-    mach_id = salt.utils.files.get_machine_identifier().get(
-        "machine_id", "no_machine_id_available"
-    )
+    mach_id = _get_machine_identifier().get("machine_id", "no_machine_id_available")
     assert repo.get_salt_working_dir() in repo._get_lock_file()
     assert repo.lock() == (
         [
